refactored Atom to abstract class added Carbon added Defect
