+++ com/sirui/analytic/driver.py
@@ -6,12 +6,12 @@ Tuning parameters
 # TODO change model. beta_delta_mu match analytic model
 from com.sirui.analytic.log_parser import LogParser
 from com.sirui.sim.config import Config
-import com.sirui.sim.TwoDSim as TwoDSim
+import com.sirui.sim.simulate as sim
 import shutil
 import os
 import logging
 
-delta_mu = [0.0]
+delta_mu = [0.2, 0.3]
 Config.time_limit = 20
 Config.SIM_TIME = None
 print('Cleaning all previous logs and images...')
@@ -21,4 +21,4 @@ os.makedirs('logs')
 
 for x in delta_mu:
         for i in range(1):
-            TwoDSim.main(x, i, logging.WARNING)
+            sim.main(x, i, logging.WARNING)
