--- com/sirui/sim/monitor.py
@@ -1,3 +1,0 @@
-class Monitor(object):
-    def __init__(self, sim_time, num_atom):
-        self.data = [[-1 for i in range(num_atom)] for j in range(sim_time)]
