+++ com/sirui/sim/utility.py
@@ -1,5 +1,6 @@
 import simpy
 from com.sirui.sim.context import Context
+import logging
 class yield_guard:
     def __enter__(self, event, id):
         try:
@@ -15,3 +16,5 @@ class yield_guard:
 
     def __exit__(self, type, value, traceback):
         pass
+
+
